scripts/requirements.txt
# pip install -r scripts/requirements.txt  OR  pip install ./scripts
httpx>=0.27
feedparser>=6.0.11
pydantic>=2.6
python-dateutil>=2.9
jinja2>=3.1
openai>=1.0
pyyaml>=6.0
markdown-it-py>=3.0
numpy>=1.26
bashlex>=0.16
Pillow>=10.0
defusedxml>=0.7
google-genai>=0.2
